Preset all select options for tabular scripts
