--- src/column_filtering.tsx
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018- The Pixie Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-import { SelectableValue } from '@grafana/data';
-
-export function getColumnsScript(
-  chosenOptions: Array<SelectableValue<number>>,
-  allColumnOptions: Array<SelectableValue<number>>
-): string {
-  // Setting script default to all the columns in the script
-  let script = allColumnOptions.map(({ label }) => `'${label}'`).join();
-
-  // Updating the script if the user selected columns to filter
-  if (chosenOptions?.length > 0) {
-    script = chosenOptions.map(({ label }) => `'${label}'`).join();
-  }
-
-  return script;
-}
